+++ smallhands/__init__.py
@@ -1,0 +1,3 @@
+from config import SmallhandsConfig
+from listener import SmallhandsListener
+from main import Smallhands
